dlong and hlong types added in ellipticTri2D and parAlmond. Need to stich in gslib types
Former-commit-id: 9fe7a40a07010fc3a9b2f8e46a8de805631f6e6b

+++ src/gsParallelGatherScatterSetup.c
@@ -7,8 +7,8 @@
 
 #include "gslib.h"
 
-void *gsParallelGatherScatterSetup(long long int NuniqueBases,
-                        				   long long int *gatherGlobalNodes,
+void *gsParallelGatherScatterSetup(int NuniqueBases,
+                        				   int *gatherGlobalNodes,
                                    int verbose){
                           
   /* gslib stuff */
